Enforce correct AsyncContext use and name "thread" (#1662)
`TestExecutor` can improve the testing of the code under test by
requiring that tasks it runs correctly handle saving and restoring of
the `AsyncContext`. It would also be useful to set the thread name
so that it can be captured by the tasks and to make the execution
context clear for logging or debugging.
Modifications:
Before any task is run the active (if any) `AsyncContext` is saved and
a special hostile instance is substituted that detects improper
access. The name of the executing thread is also adjusted while running
the task to identify the `TestExecutor` instance.
Result:
Additional testing capability for `TestExecutor` users.