Convert "ASCII-art" arrows to unicode arrows (#1668)
Motivation:
The javadoc for quite a few methods includes use of `=&gt;` to
approximate an arrow for diagrams in small figures. Converting these to
unicode characters is cleaner, makes the source the same as the
rendered output and is more accessible.
Modifications:
Use of `=&gt;` is replaced with `⇒`
Result:
Cleaner source.