Fix H2 Server with SNI bug (#1665)
Motivation:
H2 server with SNI enabled current does not work. Netty's SniHandler
does not queue writes in the same way that the SslHandler does. The h2
handler is added and writes a settings frame immediately on the server,
which is written to the socket before the handshake is completed. This
results in a handshake failure.

Modifications:
- Defer adding h2 protocol handlers to the pipeline until after the SNI
  selection is made.

Result:
H2 Server with SNI now works as expected.