`ClientEffectiveStrategyTest` and `ServerEffectiveStrategyTest` cleanups (#1667)
* `ClientEffectiveStrategyTest` cleanups
Motivation:
The parameterization pattern used for `ClientEffectiveStrategyTest`
makes it difficult to understand the configuration under test because
it is built imperatively rather than as a set of declarative options.
Some refactoring
Modifications:
Restructured test parameterization. One specific case is improved; for
the cases where no offloading to provided executor is to be performed
the test now explicitly ensures that the test fails if the executor is
used.
Result:
Easier to understand test with easier to understand test cases.