Write of HTTP/2 message may never complete
Motivation:

In #1518 we optimized HTTP/2 to write less frames when it's known that
the message body is empty. However, that change and the #1644 follow-up
for `AbstractH2DuplexHandler` incorrectly handle end of outbound stream.
`AbstractH2DuplexHandler` writes the `endStream` flag and waits for an
empty trailers. Netty h2 codec will close the stream as soon as it sees
`endStream` for request and response and will invoke
`WriteStreamSubscriber#channelClosed(...)`. If offloading is enabled,
message body can be cancelled before trailers reached the flush
strategy. As the result, `flushOnEnd` will never be triggered and some
tests hang waiting for the response that will never be flushed. It also
affects tests related to `TransportObserver` because `writeComplete`
may never be triggered.

Modifications:

- Enhance `AbstractStreamingHttpConnection` and `NettyHttpServer` to
use only meta-data when it creates a flat stream of HTTP message if the
message body is empty;
- `AbstractH2DuplexHandler`: invoke
`CloseHandler#protocolPayloadEndOutbound(...)` before writing a frame
with `endStream` flag;
- Enhance tests for `AbstractH2DuplexHandler` to verify new behavior;

Result:

Fixes #1663.