Support controlling peerHost and SNI hostname inference for GRPC clients (#1678)
Motivation:

This change follows from [a change to HTTP client builders](https://github.com/apple/servicetalk/pull/1561)
and exposes the same capabilities to GRPC client builders.

Modifications:

- `SingleAddressGrpcClientBuilder` has new methods: `inferPeerHost`,
 `inferPeerPort`, and `inferSniHostname` which allow configuring the
 behaviour of the TLS handshake and inference of the corresponding
 configurations.

Result:

Users are able to disable peerHost and peerPort inference and can now
use an empty peerHost when they so desire. The SNI hostname inference
from the provided address can also be disabled in cases when the
address used for connection should not be passed in the SNI extension.