`AllWritesPromise` should not be writable after channel is closed (#1696)
Motivation:

`AllWritesPromise` is closed when a channels becomes inactive or the
output side of the channel is shutdown. We should not try to write
more messages after that, even if a subscriber is not terminated yet.

Modifications:

- Consider `AllWritesPromise` as non-writable if any of the `state` flag
is set, including `CLOSE_OUTBOUND_ON_SUBSCRIBER_TERMINATION`;
- Add a test to verify further writes are discarded after
`WSS#channelClosed` is invoked;

Result:

No failed writes on netty pipeline after the channel is closed.