Fix flaky `HttpRequestDecoderTest.unexpectedContentAfterNoContentHeaders()`
Motivation:

`HttpRequestDecoderTest.unexpectedContentAfterNoContentHeaders()` uses
random bytes for an unexpected content. If the first two random bytes
are an upper-case letter followed by a whitespace, `HttpObjectDecoder`
may consider it as a legit HTTP method name (if there is no CRLF) and
won't throw a `DecoderException`.

Modifications:
- Use a non-random content that can not be considered as a valid input
for the next request;
- Make some internal methods `static`, where possible;

Result:

Fixes #1694.