Simplify conditional offloading (#1723)
Motivation:
The conditional offloading API of `publishOn()` and `subscribeOn()`
use of a `Supplier<? extends BooleanSupplier>` to be invoked once per
subscribe can be instead used more conventionally with a simply a
`BooleanSupplier`. If per-subscribe `BooleanSupplier` is needed it can
be provided instead in a defer() operator.
Modifications:
Replaces `Supplier<? extends BooleanSupplier>` with `BooleanSupplier`
for conditional `publishOn` and `subscribeOn`.
Ensures that the source is cancelled if the `BooleanSupplier `throws
an exception.
Result:
Simpler conditional offload API. More consistent cancel behavior.