`MultiAddressHttpClientBuilder` deprecation and de-deprecation refactoring (#1722)
MultiAddressHttpClientBuilder deprecation and de-deprecation refactoring

Motivation:

The `MultiAddressHttpClientBuilder` deprecated methods which should stay
and allow configuring the `ExecutionContext` (`ioExecutor,
`executionStrategy`, `bufferAllocator`). Furthermore, methods
influencing the client filters have not been deprecated, but should, as
the `SingleAddressInitializer` allows configuring the filters in a
satisfactory fashion.

Modifications:

- Removed `@Deprecated` annotation from `ioExecutor`, `executionStrategy`,
  and `bufferAllocator` methods,
- Deprecated `MultiAddressHttpClientFilterFactory`,
- Deprecated `appendClientFilter` methods accepting
  `MultiAddressHttpClientFilterFactory`,
- Deprecated `StreamingHttpClientFilterFactory#asMultiAddressClientFilter`.

Result:

The current API reflects the future direction for
`MultiAddressHttpClientBuilder` and how it should be used.