Propogate InterruptedException to the NettyPipelinedConnectionTest tests
Motivation:

The current usage of awaitCancelledUninterruptibly() catches
InterruptedException and silently discards it. jUnit5 uses
thread interrupt to stop tests after the timeout expires. As a
result, these tests may hang forever as the InterruptedException
is not propogated.

Modifications:

- use awaitCancelled(), which throws InterruptedException
- Declare InterruptedException on method declaration.

Result:

The InterruptedException is propogated to the test methods,
which allows jUnit5 to stop a hanging test when the test's
timeout expires.