Improve split cookie crumbs test in `H2ToStH1UtilsTest`
Motivation:

We used `ConstantHashCharSequence` to make sure that two headers fall
into the same bucket in internal `HttpHeaders` data structure. However,
the `hashCode()` is taken only if `CharSequence` implements
`AsciiBuffer` or `AsciiString`. Otherwise, a static utility is used
to compute case-insensitive `hashCode`. As the result,
`ConstantHashCharSequence` does not provide what we expected.

Modifications:

- Remove `ConstantHashCharSequence` and use `caseInsensitiveHashCode` /
`hashCodeAscii` utilities to make sure that two header names result
in the same bucket based on their `hashCode` value and `ARRAY_SIZE_HINT`

Result:

Test verifies the behavior it was created for.