Simplify API hierarchy of HTTP and gRPC client builders (#1743)
Simplify API hierarchy of HTTP and gRPC client builders

Motivation:

The current hierarchy of HTTP and gRPC client builders is complicated
and lacks alignment between both stacks. The HTTP hierarchy in
particular carries along some historical context which was recently
updated to make the `Partitioned`- and `MultiAddress`- `HttpClientBuilder`s
independent of the API useful for `SingleAddressHttpClientBuilder` as
keeping them consistent was problematic and led to the introduction of
`SingleAddressInitializer` mechanism in both cases. After recent
untangling of particular clients' dependency on those shared APIs it's
now time to simplify the hierarchy, which this change attempts.

Modifications:

- Removed `BaseGrpcClientBuilder`,
- Removed `SingleAddressGrpcClientBuilder`,
- Removed `BaseSingleAddressHttpClientBuilder`,
- Moved methods of `HttpClientBuilder` into
  `SingleAddressGrpcClientBuilder` to highlight their meaning in the
  context of a client that is dedicated to a particular address. This
  change left the `HttpClientBuilder` interface empty,
- Merged `ExecutionContextAwareHttpBuilder` and
  `HttpClientBuildFinalizer` into `HttpClientBuilder` as those
  package-private interfaces were useful in the untangling process. Now
  they're not required any more and can be incorporated into
  `HttpClientBuilder` which solely represents the responsibilities that
  those interfaces encapsulated,
- Reverted `MultiAddressHttpClientBuilder` and
  `PartitionedHttpClientBuilder` to abstract classes instead of
  interfaces to remain consistent with the server builders and the gRPC
  counterparts.

Result:

The API hierarchy is simplified, uses similar concepts for different
levels across the HTTP and gRPC stack and the more complex multi- and
partitioned- client builders have a common base interface that doesn't
carry unnecessary mechanisms from the single address use case. At the
same time, for most usages where the `HttpClients` and `GrpcClients`
factories are used, there would be no action required from the users.