Support additional constant delay for retries
Motivation:

Some errors may require a constant delay before they can be retried.
Example of that can be an unavailable service that responds with a retry-after: period
header. This period can be used as a constant delay when retrying.
On top of the constant delay, retry-strategies configured for a retry operator
will still need to be taken into consideration to offer additional jitter to avoid
retry storms at the same time.

Modifications:

Introduced a DelayedRetry marker interface that exposes a duration API. That in
turn is consulted inside the retry filter to compose the total delay duration.

Result:

Retry filter supports additionally constant delay if the error contains that extra info.