ZipkinPublisher remove unnecessary data
Motivation:
We currently publish a time stamp for begin, a duration, and a timestamp
for end. We can remove the end timestamp which can be derived by doing
begin+end. We also publish a tag for the type of span, but there is also
an explicit field for the type.

Modifications:
- Remove the tag from the span before publishing, remove end timestamp
  in ZipkinPublisher.

Result:
Less unnecessary data published from ZipkinPublisher.