gRPC add MethodDescriptor Collection to services (#1764)
Motivation:
It is currently not possible for a user to get a Collection of
MethodDescriptors corresponding to a service. The collection of
MethodDescriptors is necessary to build dynamic routing in the event of
a custom transport.

Modifications:
- Add `Collection<MethodDescriptor<?, ?>>` method to
  `GrpcBindableService` API.
- Add a unit test which introduces a custom in memory transport and
  dynamic routing for a service, and also a client that routes to that
  transport in memory.

Result:
It is now possible to get a Collection of MethodDescriptor from a
service, and dynamically build routing for a custom transport.