Add missing subscribeShareContext to missing static sources
Motivation:
The subscribeShareContext was added in a few cases when defer was used
but not added to all return paths. It should also be added to static
sources like succeed and failed for consistency.

Modifications:
- Update deprecated ContentCoding filters to use subscribeShareContext
  on static succeeded operators.
- Move subscribeShareContext in tracing filters to the usage of defer,
  to make the need for subscribeShareContext more clear.

Result:
More consistent usage of subscribeShareContext on static sources when
control flow otherwise expects sharing context.