Use bom to determine modules to compare (#1821)
Motivation:
The japicmp script used the local directory to determine the modules to
compare, even if that local directory had nothing to do with the
versions being compared.
Modifications:
Use the bom files from the versions being compared to determine which
modules to compare. The local directory will be used rather than bom
file if doing a local compare. The modules compared is the intersection
of the modules present in both versions.
Result:
Smarter comparison