NettyHttpServerTest.testErrorBeforeRead handle exception
Motivation:
NettyHttpServerTest.testErrorBeforeRead has the server intentionally
write an error without first consuming the request content. The server
may close the connection before the client is done writting and if that
is the case a ClosedChannelException may be observed and fail the test.

Modifications:
- Allow ClosedChannelException when making a request

Result:
Fixes https://github.com/apple/servicetalk/issues/245