Identify `LoadBalancer` instance using a `targetResource(String)` (#1831)
Motivation:

An instance of `LoadBalancer` will do connection selection over a
collection of hosts provided by a `Publisher` of resolved addresses.
While debugging issues it helps to identify the intended resource that
the `LoadBalancer` is helping to reach. Unfortunately, at it's level,
the unresolved address is unknown and it may be the case that the
collection of hosts represents more than a single unresolved address. To
make the debugging process more insightful, a `String` parameter has
been added, named `targetResource` to identify a particular instance.

Modifications:

- Added a `String` parameter to `LoadBalancerFactory`'s
  `newLoadBalancer` method: `targetResource` in the form of a new
  default method in the interface to avoid breaking the API and the
  contract of `FunctionalInterface`,
- Using the address from
  `io.servicetalk.http.netty.DefaultSingleAddressHttpClientBuilder.HttpClientBuildContext`
  when creating instances of `LoadBalancer` in the
  DefaultSingleAddressHttpClientBuilder`.

Result:

Debugging connectivity issues originating from unhealthy state of
hosts behind the `LoadBalancer` should be easier as it is now possible
to identify the target resource that the `LoadBalancer` is connecting
to.