Deprecate `AwaitUtils#pollUninterruptibly`
Motivation:
The current usage of AwaitUtils catches
InterruptedException and silently discards it. jUnit5 uses
thread interrupt to stop tests after the timeout expires. As a
result, these tests may hang forever as the InterruptedException
is not propagated.

Modifications:

- introduced `AwaitUtils#poll`

Result:
`AwaitUtils#poll` now throws unchecked exception
in the case that the thread is interrupted. Meaning that
JUnit5 will be able to stop the tests by interrupting
when it timeouts.