Support proxy for `MultiAddressHttpClientBuilder` and for clients using already resolved address (#1751)
Motivation:

Using `MultiAddressHttpClientBuilder` it's currently not possible to set a proxy. Users of this API should also be able to specify a proxy for each address that the client connects to.

Modifications:

* Added `SingleAddressHttpClientBuilder#proxyAddress` to allow specifying proxy on the builder,
* Simplified internal logic of mapping specified host to a `CharSequence` in `DefaultSingleAddressHttpClientBuilder`,
* Introduced a test for the feature: `HttpMultiProxyTest`,
* Added a test for builder reuse: `HttpProxyTest#testBuilderReuseEachClientUsesOwnProxy`,
* Deprecated  *`viaProxy` methods in `HttpClients`,
* Added proxy support for resolved address client (configured via `HttpClients#forResolvedAddress`*) - proxy did not work with `forResolvedAddressViaProxy` variants,
* Improved Javadocs for resolved address use.

Result:

It's possible to specify a proxy for each individual address that clients created using `MultiAddressHttpClientBuilder` can reach. Using proxy with a resolved address client is also now possible. The API got a bit simpler too.