Handle InterruptedException in tests consistently Motivation: There are some test files where thread's interrupted flag is not reset after InterruptedException is caught, as well as where InterruptedException is being wrapped into RuntimeException, which JUnit5 may not recognize, when stopping the tests after timeout.
Modifications:

- Use `Thread.currentThread().interrupt();` where it was forgotten
- Use unchecked exception with `throwException` instead of `RuntimeException`
- Where possibly simply rethrow the caught exception.

Result:
That did not work out as nicely as expected. Some tests are failing
locally, which I am still trying to figure out how to fix.

some questions:
- how to handle this in HttpOffloadingTest.java:345?
I dont really see how the errors in errors queue is handled, so had
no idea if there is anything I could fix here.

- Is it okay how I handle InterruptedException in `TestPublisherSubscriber#takeOnNext(int n)`?
as I understand if the exception was thrown we should propagate it at
some point, but wouldn't that break the current method?

That's about it, if I forgot something I will add it as a comment
to the github.