Fix `DefaultGrpcClientBuilder`s internal direct initializer.
Motivation:

Recent deprecation of methods in `GrpcClientBuilder` required an
introduction of a helper `HttpInitializer` for methods previously
targetting the internal instance of underlying
`SingleAddressHttpClientBuilder`. The `append` method was usued
incorrectly as the enriched initializer was discarded instead of
assigned to the state. This change fixes that deprecated behaviour.

Modifications:

Assign the composite initializer to the state of the builder instead of
discarding the modified version.

Result:

Proper state management of the `DefaultGrpcClientBuilder`.