Enhance `japicmp.sh` to be less strict on finding artifacts
Motivation:

When we are comparing with a local build, we should search only for
local artifacts, no need to check old bom. Some local folders may also
start with `servicetalk-` prefix, but may be without artifacts. If an
artifact is not found, log and continue to the next one.

Modifications:

- Find `OLD_ARTIFACTS` only for non-local builds;
- Continue if any artifact is not found;
- Make `BUILD_DIR` parametrized;

Result:

Comparison with a local build focused only on local artifacts, script
doesn't fail if an artifact not found.