Remove gRPC filters
Motivation:

gRPC filters are not as flexible as HTTP client filters. Their
reuse is limited due to the strict typing with gRPC definitions. HTTP
filters are the best choice most of the time. If required, users can
embed their custom logic in the actual client implementation where the
decoded protos can be accessed. Recently the filter related classes and
methods have been deprecated. The intention is to remove them in the
0.42 release.

Modifications:

- Adjusted generic types in return values and parameters to not include
  filters (breaking change),
- Removed deprecated `GrpcClientFilterFactory`,
- Removed deprecated `GrpcClientFactory#appendClientFilter`,
- Removed `GrpcClientFactory#appendClientFilterFactory`, and
  `GrpcClientFactory#newFilter` methods,
- Removed code from the `Generator` class related to filters,
- Adjusted tests to not exercise logic that relies on or tests filters.

Result:

gRPC filters have been removed.