Deprecate `ExecutionContext` related methods in `GrpcServerBuilder`
Motivation:

Recent changes to `GrpcServerBuilder` attempted decoupling the gRPC
layer from the underlying HTTP transport in terms of the builder
APIs. Methods that serve the purpose of configuring the
`ExecutionContext` have not been deprecated due to the way the actual
`ServerContext` is created with regards to `GrpcRoutes`. This change
deprecates four methods and introduces a mechanism to transport the
underlying `HttpServerBuilder`'s `ExecutionContext` to populate the
context in routes.

Modifications:

- Deprecated `ioExecutor`, `executor`, `bufferAllocator`, and
  `executionStrategy` methods in `GrpcServerBuilder`,
- Introduced an internal `ExecutionContextInterceptorHttpServerBuilder`
  to capture the configured `ExecutionContext` via the
  `GrpcServerBuilder.HttpInitializer`,
- Adjusted test cases to use the new APIs.

Result:

All of the `GrpcServerBuilder` methods that served as a facade to
`HttpServerBuilder` have now been deprecated and there is a single way
to configure the underlying transport using
`initializeHttp(HttpServerBuilder)` method.