Ensure `asyncClose` `Completable` is offloaded (#1900)
Motivation:
The terminal signal for the `Completable` returned by `asyncClose` or
`asyncCloseGracefully` should be offloaded.
Modifications:
Re-order the steps of the `CompositeCloseable` so that the offloaded
onClose is the source of the terminal signal. Prior to #1585 the order
did not matter as all of the `Completable` were offloaded.
Result:
`Completable` result of `HttpServer.closeAsyncGracefully()` no longer
invokes `onComplete()` on the IoThread.