Adjusting `ServiceDiscovererEvent` contract
Motivation:

Current API of `ServiceDiscovererEvent` is capable of representing the
boolean status of the host being either available or not available. As
seen in the implementation of RoundRobinLoadBalancer`, the resulting
states from the Service Discovery system can have more meanings than a
simple boolean. To prepare for enhanecements, the contract has been
adjusted to allow more states and make it possible for further additions
if those are needed by particular implementations.

Modifications:

- Deprecated `ServiceDiscovererEvent#isAvailable` method and added a
  default implementation to allow removal in child classes,
- Added `ServiceDiscoveryStatus` class,
- Added `ServiceDiscovererEvent#status` method with default
  implementation to prevent breaking the existing implementations,
- Replaced existing uses of `isAvailable` with comparisons of `status`
  against the constants in `ServiceDiscoveryStatus`,
- Added TODOs in order to reflect the need to adapt the codebase to
  support non-binary statuses.

Result:

`ServiceDiscovererEvent`s are prepared for more advanced, non-binary
statuses. An example is the now introduced `EXPIRED` event, which can have a
different meaning than `UNAVAILABLE`. If the `ServiceDiscoverer` decides
to emit such a status, the subscribers can provide additional
functionality in that case.