Motivation: (#1916)
Several Netty components are used as runtime dependencies by the
`tcp-netty-internal` test fixtures. The current Maven POM does not
include versions for these components.
Modifications:
The Netty BOM is imported as a platform to ensure that the versions are
available in the published Maven POM file for this module.
Result:
Maven builds including `servicetalk-tcp-netty-internal` module have the
versions they need.