Enhance javadoc to clarify StackOverflowError possibility (#1925)
Motivation:
Publisher [repeat | retry] operators may cause StackOverflowError if too
many recovery operations are done consecutively on the same thread. This
condition can be difficult to detect as the onError recovery path is
unlikely to be called. We should highlight this possibility in the
javadocs for these methods.

Modifications:
- Publisher, Single, Completable repeat* and retry* operators have
  javadoc that clarifies the risk of StackOverflowError and mitigation
  technique.

Result:
Javadocs highlight risks of StackOverflowError and how it can be
avoided.