MalformedDataAfterHttpMessageTest.afterResponseNextClientRequestSucceeds fix (#1926)
Motivation:
afterResponseNextClientRequestSucceeds attempts to validate that
NettyChannelPublisher preemptively closes the channel after request
processing concluded if there is an error pending to prevent subsequent
requests from using this connection. This condition may not be hit in
all scenarios due to sequencing of offloading as well as when bytes are
read from the network and the error condition is observed. This results
in the test being flaky and failing.

Modifications:
- Make afterResponseNextClientRequestSucceeds more robust by inserting a
  RetryingHttpRequesterFilter and adjusting the assertions to expect
  retry may happen.

Result:
Fixes https://github.com/apple/servicetalk/issues/1889.