`SingleThreadedScheduler` should fallback execution upon offloading failure
Motivation:

When the `Executor` provided by the user fails to accept the task
executed by the `SingleThreadedScheduler`, the exception is ignored and
the scheduler is no longer usable. The exceptions from the user land
should not surface to the global scheduler and the implementation should
do its' best to run the task even at the cost of skipping offloading.

Modification:

`DefaultExecutor.SingleThreadedScheduler#schedule` is guarded against
rejections from the user's `offloadExecutor` and triest to fallback
and execute the task on its' own thread.

Result:

User error does not prevent the `SingleThreadedScheduler` from executing
future tasks.