Give visibility to the assigned HTTP/2 Stream ID (#1969)
Motivation:

`HttpServerContext` always has an HTTP/2 stream id as part of the
`ctx.toString()`. On the client-side, the identifier is assigned only
after the first write of `HEADERS` frame is made. We can give visibility
into the assigned stream id through the `StreamObserver`.

Modifications:

- Add `StreamObserver#streamIdAssigned(long)` callback;
- Invoke the new callback when the steam id is known;
- Test new behavior;

Result:

Users have visibility into the assigned HTTP/2 stream id and when it's
assigned.